# FoodBook/FoodBook/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FoodBookAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Ingredients": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientsForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientsForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientsForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ingredients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ingredients/{id}": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientsForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IngredientsForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "kcal": {
            "type": "number",
            "format": "float"
          },
          "fats": {
            "type": "number",
            "format": "float"
          },
          "proteins": {
            "type": "number",
            "format": "float"
          },
          "salt": {
            "type": "number",
            "format": "float"
          },
          "sugar": {
            "type": "number",
            "format": "float"
          },
          "carbs": {
            "type": "number",
            "format": "float"
          },
          "amountOfUnit": {
            "type": "number",
            "format": "float"
          },
          "idUnitOfMeasurement": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unitTitle": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}